requirements.txt
pyusb
psutil